test: create unit tests
