+++ apps/meteor/app/lib/server/functions/notifications/messageContainsHighlight.ts
@@ -1,0 +1,21 @@
+import type { IMessage } from '@rocket.chat/core-typings';
+import { escapeRegExp } from '@rocket.chat/string-helpers';
+
+/**
+ * Checks if a message contains a user highlight
+ *
+ * @param {string} message
+ * @param {array|undefined} highlights
+ *
+ * @returns {boolean}
+ */
+export function messageContainsHighlight(message: Pick<IMessage, 'msg'>, highlights: string[] | undefined): boolean {
+	if (!highlights || highlights.length === 0) {
+		return false;
+	}
+
+	return highlights.some((highlight: string) => {
+		const regexp = new RegExp(`\\b${escapeRegExp(highlight)}\\b`, 'i');
+		return regexp.test(message.msg);
+	});
+}
